Refactor devnet and cast types, move data and serializers to respective packages
