+++ lib/src/test/kotlin/starknet/utils/data/SnCastTypes.kt
@@ -1,11 +1,10 @@
-package starknet.utils
+package starknet.utils.data
 
 import com.swmansion.starknet.data.types.Felt
-import com.swmansion.starknet.data.types.transactions.*
-import kotlinx.serialization.DeserializationStrategy
 import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.json.*
+import starknet.utils.data.serializers.SnCastResponsePolymorphicSerializer
 
 @OptIn(ExperimentalSerializationApi::class)
 @Serializable
@@ -30,29 +29,6 @@ enum class SnCastCommand {
 sealed class SnCastResponse {
     abstract val command: SnCastCommand
     abstract val error: String?
-}
-
-internal object SnCastResponsePolymorphicSerializer : JsonContentPolymorphicSerializer<SnCastResponse>(SnCastResponse::class) {
-    override fun selectDeserializer(element: JsonElement): DeserializationStrategy<out SnCastResponse> {
-        val jsonObject = element.jsonObject
-
-        val commandObject = jsonObject.getOrElse("command") { throw IllegalArgumentException("Missing command type in sncast response") }
-        val command = Json.decodeFromJsonElement(SnCastCommand.serializer(), commandObject)
-
-        val error = jsonObject["error"]?.jsonPrimitive?.content
-        error?.let {
-            throw SnCastCommandFailed(commandObject.jsonPrimitive.content, error)
-        }
-
-        return when (command) {
-            SnCastCommand.ACCOUNT_CREATE -> AccountCreateSnCastResponse.serializer()
-            SnCastCommand.ACCOUNT_DEPLOY -> AccountDeploySnCastResponse.serializer()
-            SnCastCommand.DECLARE -> DeclareSnCastResponse.serializer()
-            SnCastCommand.DEPLOY -> DeploySnCastResponse.serializer()
-            SnCastCommand.INVOKE -> InvokeSnCastResponse.serializer()
-            else -> throw IllegalArgumentException("Invalid command type")
-        }
-    }
 }
 
 @OptIn(ExperimentalSerializationApi::class)
